allwinner: Add functions to control CPU power/reset
sun50i_cpu_on will be used by the PSCI implementation to initialize
secondary cores for SMP. Unfortunately, sun50i_cpu_off is not usable by
PSCI directly, because it is not possible for a CPU to use this function
to power itself down. (Power cannot be shut off until the outputs are
clamped, and MMIO does not work once the outputs are clamped.)
But at least CPU0 can shutdown the other cores early in the BL31 boot
process and before shutting down the system.

[Andre: some cleanups]

Change-Id: Id79e40cb6a278d0bd51f98453b11af9c4db4784b
Signed-off-by: Samuel Holland <samuel@sholland.org>
Signed-off-by: Andre Przywara <andre.przywara@arm.com>