allwinner: Add Allwinner H6 SoC support
The Allwinner H6 is a new SoC. It shares most peripherals with the other
ARMv8 Allwinner SoCs (A64/H5), but has a completely different memory
map.
Introduce a separate platform target, which includes a different header
file to cater for the address differences.
The ATF platform name is "sun50i_h6".

Signed-off-by: Andre Przywara <andre.przywara@arm.com>